+++ client/common/media/icons/Overview.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const DashboardIcon = props => {
+const OverviewIcon = props => {
 	return (
 		<svg {...props} fill="none" stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
 			<rect x="3" y="3" width="18" height="18" rx="2" ry="2" />
@@ -10,4 +10,4 @@ const DashboardIcon = props => {
 	);
 };
 
-export default DashboardIcon;
+export default OverviewIcon;